# common-platform-config-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-config-mgmt
  description: |
    Common Platform Config Management Core Application

    ## UUID Strategy

    This API uses UUIDs (Universally Unique Identifiers) for all entity primary keys and foreign key relationships. This design provides several benefits:

    - **Global Uniqueness**: UUIDs are globally unique, eliminating ID conflicts when merging data from different environments
    - **Distributed Systems Support**: No need for centralized sequence coordination across multiple database instances
    - **Enhanced Security**: Non-sequential IDs make it harder to guess valid identifiers
    - **Better Distribution**: UUIDs provide better distribution in sharded or partitioned databases

    ### ID Field Format

    All entity ID fields (primary keys and foreign keys) use the following format:
    - **Type**: `string`
    - **Format**: `uuid`
    - **Example**: `550e8400-e29b-41d4-a716-446655440000`

    ### Exception: Country IDs

    Country IDs are the only exception and use sequential integer identifiers for simplicity and performance in geographical lookups.

    ### API Usage

    When making API requests:
    - Always provide UUIDs in the standard format (8-4-4-4-12 hexadecimal digits)
    - The API will validate UUID format and return appropriate error messages for invalid UUIDs
    - All responses will include UUIDs in the standard string format
  contact:
    name: Firefly Software Solutions Inc
...
  version: 1.0.0
servers:
...
paths:
...
  /api/v1/providers/{id}:
    get:
      tags:
        - Providers
      summary: Get a provider by ID
      description: Returns a provider based on the ID
      operationId: getProviderById
      parameters:
        - name: id
          in: path
          description: ID of the provider to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProviderDTO'
        '404':
          description: Provider not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProviderDTO'
    put:
      tags:
        - Providers
      summary: Update an existing provider
      description: Updates an existing provider and returns it
      operationId: updateProvider
      parameters:
        - name: id
          in: path
          description: ID of the provider to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderDTO'
        required: true
      responses:
        '200':
          description: Provider updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProviderDTO'
        '400':
          description: Invalid input
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProviderDTO'
        '404':
          description: Provider not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProviderDTO'
    delete:
      tags:
        - Providers
      summary: Delete a provider
      description: Deletes a provider
      operationId: deleteProvider
      parameters:
        - name: id
          in: path
          description: ID of the provider to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Provider deleted
        '404':
          description: Provider not found
...
  /api/v1/provider-contracts/{id}:
    get:
      tags:
        - Provider Contracts
      summary: Get a provider contract by ID
      description: Returns a provider contract based on the ID
      operationId: getProviderContractById
      parameters:
        - name: id
          in: path
          description: ID of the provider contract to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProviderContractDTO'
        '404':
          description: Provider contract not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProviderContractDTO'
    put:
      tags:
        - Provider Contracts
      summary: Update an existing provider contract
      description: Updates an existing provider contract and returns it
      operationId: updateProviderContract
      parameters:
        - name: id
          in: path
          description: ID of the provider contract to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderContractDTO'
        required: true
      responses:
        '200':
          description: Provider contract updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProviderContractDTO'
        '400':
          description: Invalid input
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProviderContractDTO'
        '404':
          description: Provider contract not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProviderContractDTO'
    delete:
      tags:
        - Provider Contracts
      summary: Delete a provider contract
      description: Deletes a provider contract
      operationId: deleteProviderContract
      parameters:
        - name: id
          in: path
          description: ID of the provider contract to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Provider contract deleted
        '404':
          description: Provider contract not found
  /api/v1/provider-contract-statuses/{id}:
    get:
      tags:
        - Provider Contract Statuses
      summary: Get a provider contract status by ID
      description: Returns a provider contract status based on the ID
      operationId: getProviderContractStatusById
      parameters:
        - name: id
          in: path
          description: ID of the provider contract status to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProviderContractStatusDTO'
        '404':
          description: Provider contract status not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProviderContractStatusDTO'
    put:
      tags:
        - Provider Contract Statuses
      summary: Update an existing provider contract status
      description: Updates an existing provider contract status and returns it
      operationId: updateProviderContractStatus
      parameters:
        - name: id
          in: path
          description: ID of the provider contract status to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderContractStatusDTO'
        required: true
      responses:
        '200':
          description: Provider contract status updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProviderContractStatusDTO'
        '400':
          description: Invalid input
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProviderContractStatusDTO'
        '404':
          description: Provider contract status not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProviderContractStatusDTO'
    delete:
      tags:
        - Provider Contract Statuses
      summary: Delete a provider contract status
      description: Deletes a provider contract status
      operationId: deleteProviderContractStatus
      parameters:
        - name: id
          in: path
          description: ID of the provider contract status to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Provider contract status deleted
        '404':
          description: Provider contract status not found
...
  /api/v1/providers/filter:
    post:
      tags:
        - Providers
      summary: Filter providers
      description: Returns a filtered list of providers based on criteria
      operationId: filterProviders
      parameters:
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.code
          in: query
          description: Exact filter for code
          required: false
          schema:
            type: string
        - name: filters.name
          in: query
          description: Exact filter for name
          required: false
          schema:
            type: string
        - name: filters.description
          in: query
          description: Exact filter for description
          required: false
          schema:
            type: string
        - name: filters.providerTypeId
          in: query
          description: Exact filter for provider type id
          required: false
          schema:
            type: string
        - name: filters.providerType
          in: query
          description: Exact filter for provider type
          required: false
          schema:
            type: string
        - name: filters.providerStatusId
          in: query
          description: Exact filter for provider status id
          required: false
          schema:
            type: string
        - name: filters.providerStatus
          in: query
          description: Exact filter for provider status
          required: false
          schema:
            type: string
        - name: filters.apiBaseUrl
          in: query
          description: Exact filter for api base url
          required: false
          schema:
            type: string
        - name: filters.webhookUrl
          in: query
          description: Exact filter for webhook url
          required: false
          schema:
            type: string
        - name: filters.callbackUrl
          in: query
          description: Exact filter for callback url
          required: false
          schema:
            type: string
        - name: filters.logoUrl
          in: query
          description: Exact filter for logo url
          required: false
          schema:
            type: string
        - name: filters.documentationUrl
          in: query
          description: Exact filter for documentation url
          required: false
          schema:
            type: string
        - name: filters.supportUrl
          in: query
          description: Exact filter for support url
          required: false
          schema:
            type: string
        - name: filters.contactName
          in: query
          description: Exact filter for contact name
          required: false
          schema:
            type: string
        - name: filters.contactEmail
          in: query
          description: Exact filter for contact email
          required: false
          schema:
            type: string
        - name: filters.contactPhone
          in: query
          description: Exact filter for contact phone
          required: false
          schema:
            type: string
        - name: filters.technicalContactName
          in: query
          description: Exact filter for technical contact name
          required: false
          schema:
            type: string
        - name: filters.technicalContactEmail
          in: query
          description: Exact filter for technical contact email
          required: false
          schema:
            type: string
        - name: filters.technicalContactPhone
          in: query
          description: Exact filter for technical contact phone
          required: false
          schema:
            type: string
        - name: filters.region
          in: query
          description: Exact filter for region
          required: false
          schema:
            type: string
        - name: filters.timezone
          in: query
          description: Exact filter for timezone
          required: false
          schema:
            type: string
        - name: filters.currencyCode
          in: query
          description: Exact filter for currency code
          required: false
          schema:
            type: string
        - name: filters.maxRequestsPerSecond
          in: query
          description: Exact filter for max requests per second
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[maxRequestsPerSecond].from
          in: query
          description: Filter max requests per second from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[maxRequestsPerSecond].to
          in: query
          description: Filter max requests per second to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.maxConcurrentRequests
          in: query
          description: Exact filter for max concurrent requests
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[maxConcurrentRequests].from
          in: query
          description: Filter max concurrent requests from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[maxConcurrentRequests].to
          in: query
          description: Filter max concurrent requests to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.requiresAuthentication
          in: query
          description: Exact filter for requires authentication
          required: false
          schema:
            type: boolean
        - name: filters.authenticationType
          in: query
          description: Exact filter for authentication type
          required: false
          schema:
            type: string
        - name: filters.supportsWebhooks
          in: query
          description: Exact filter for supports webhooks
          required: false
          schema:
            type: boolean
        - name: filters.supportsCallbacks
          in: query
          description: Exact filter for supports callbacks
          required: false
          schema:
            type: boolean
        - name: filters.supportsPolling
          in: query
          description: Exact filter for supports polling
          required: false
          schema:
            type: boolean
        - name: filters.pollingIntervalSeconds
          in: query
          description: Exact filter for polling interval seconds
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[pollingIntervalSeconds].from
          in: query
          description: Filter polling interval seconds from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[pollingIntervalSeconds].to
          in: query
          description: Filter polling interval seconds to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.active
          in: query
          description: Exact filter for active
          required: false
          schema:
            type: boolean
        - name: filters.version
          in: query
          description: Exact filter for version
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[version].from
          in: query
          description: Filter version from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[version].to
          in: query
          description: Filter version to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseProviderDTO'
...
  /api/v1/provider-contract-statuses:
    post:
      tags:
        - Provider Contract Statuses
      summary: Create a new provider contract status
      description: Creates a new provider contract status and returns it
      operationId: createProviderContractStatus
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderContractStatusDTO'
        required: true
      responses:
        '201':
          description: Provider contract status created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProviderContractStatusDTO'
        '400':
          description: Invalid input
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProviderContractStatusDTO'
  /api/v1/provider-contract-statuses/filter:
    post:
      tags:
        - Provider Contract Statuses
      summary: Filter provider contract statuses
      description: Returns a filtered list of provider contract statuses based on criteria
      operationId: filterProviderContractStatuses
      parameters:
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.code
          in: query
          description: Exact filter for code
          required: false
          schema:
            type: string
        - name: filters.name
          in: query
          description: Exact filter for name
          required: false
          schema:
            type: string
        - name: filters.description
          in: query
          description: Exact filter for description
          required: false
          schema:
            type: string
        - name: filters.active
          in: query
          description: Exact filter for active
          required: false
          schema:
            type: boolean
        - name: filters.version
          in: query
          description: Exact filter for version
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[version].from
          in: query
          description: Filter version from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[version].to
          in: query
          description: Filter version to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseProviderContractStatusDTO'
...
components:
  schemas:
...
    ProviderContractStatusDTO:
      required:
        - code
        - name
      type: object
      properties:
        id:
          type: string
          description: Provider contract status ID
          format: uuid
          readOnly: true
        code:
          maxLength: 50
          minLength: 2
          type: string
          description: Unique code for the provider contract status
          example: ACTIVE
        name:
          maxLength: 100
          minLength: 2
          type: string
          description: Name of the provider contract status
          example: Active
        description:
          type: string
          description: Description of the provider contract status
          example: Provider contract is active and operational
        active:
          type: boolean
          description: Whether the provider contract status is active
          default: true
        version:
          type: integer
          description: Version for optimistic locking
          format: int64
          readOnly: true
        createdAt:
          type: string
          description: Creation timestamp
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          description: Last update timestamp
          format: date-time
          readOnly: true
      description: Provider Contract Status information
...